requirements.txt

# =============================================
# AquaTrace Python Dependencies
# Install with: pip install -r requirements.txt
# =============================================

# Flask Framework
flask==3.0.0
flask-sqlalchemy==3.1.1
werkzeug==3.0.1

# Environment Variables
python-dotenv==1.0.0

# Serial Communication (for hardware sensors)
pyserial==3.5

# SMS Alerts (Optional - for Twilio integration)
twilio==8.10.0

# Data Processing (Optional - for advanced features)
pandas==2.1.4
numpy==1.26.2

# Plotting (Optional - for data visualization)
matplotlib==3.8.2

# Database Migrations (Optional)
flask-migrate==4.0.5

# CSRF Protection (Recommended for production)
flask-wtf==1.2.1

# Testing
pytest==7.4.3
pytest-flask==1.3.0

# Code Quality (Development)
black==23.12.1
flake8==6.1.0
pylint==3.0.3

# Production Server
gunicorn==21.2.0

# =============================================
# Installation Notes:
# =============================================
# 1. Create virtual environment: python -m venv venv
# 2. Activate:
#    - Windows: venv\Scripts\activate
#    - Linux/Mac: source venv/bin/activate
# 3. Install: pip install -r requirements.txt
# 4. Upgrade pip if needed: pip install --upgrade pip
# =============================================